+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # ExAcme Changelog
 
-## Unreleased
+## 0.4.0 (2025-03-02)
 
 - Added support for handling external account binding (https://datatracker.ietf.org/doc/html/rfc8555#section-7.3.4)
 - Added ZeroSSL as a named directory URL.